requirements.txt
pyyaml
pyftpdlib